Fix `i128` error codes

+++ sui/sources/i128.move
@@ -20,10 +20,10 @@ module movemate::i128 {
     const GREATER_THAN: u8 = 2;
 
     /// @dev When trying to convert from a u128 > MAX_I128_AS_U128 to an I128.
-    const ECONVERSION_FROM_U128_OVERFLOW: u128 = 0;
+    const ECONVERSION_FROM_U128_OVERFLOW: u64 = 0;
 
     /// @dev When trying to convert from an negative I128 to a u128.
-    const ECONVERSION_TO_U128_UNDERFLOW: u128 = 1;
+    const ECONVERSION_TO_U128_UNDERFLOW: u64 = 1;
 
     /// @notice Struct representing a signed 128-bit integer.
     struct I128 has copy, drop, store {
